♻️ 계정 초기 설정 폼 교체
